requirements.txt
streamlit==1.20.0
google-generativeai==0.2.0
PyPDF2==3.0.0
python-docx==0.8.11
